+++ cmake.config/config.h.in
@@ -18,7 +18,6 @@
 #cmakedefine HAVE_FD_CLOEXEC
 #cmakedefine HAVE_FSEEKO
 #cmakedefine HAVE_LANGINFO_H
-#cmakedefine HAVE_LOCALE_H
 #cmakedefine HAVE_NL_LANGINFO_CODESET
 #cmakedefine HAVE_NL_MSG_CAT_CNTR
 #cmakedefine HAVE_PWD_FUNCS
