+++ src/nvim/generators/gen_ex_cmds.lua
@@ -66,7 +66,6 @@ defsfile:write(string.format([[
 #include "nvim/ex_session.h"
 #include "nvim/help.h"
 #include "nvim/indent.h"
-#include "nvim/locale.h"
 #include "nvim/lua/executor.h"
 #include "nvim/mapping.h"
 #include "nvim/mark.h"
@@ -75,6 +74,7 @@ defsfile:write(string.format([[
 #include "nvim/message.h"
 #include "nvim/ops.h"
 #include "nvim/option.h"
+#include "nvim/os/lang.h"
 #include "nvim/profile.h"
 #include "nvim/quickfix.h"
 #include "nvim/runtime.h"
