+++ src/nvim/os/lang.h
@@ -1,6 +1,9 @@
 #ifndef NVIM_OS_LANG_H
 #define NVIM_OS_LANG_H
 
+#include "nvim/ex_cmds_defs.h"
+#include "nvim/types.h"
+
 #ifdef INCLUDE_GENERATED_DECLARATIONS
 # include "os/lang.h.generated.h"
 #endif
